requirements.txt
# Core dependencies
streamlit>=1.30.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Data manipulation and visualization
pandas>=2.1.0
plotly>=5.18.0

# Optional enhancements
streamlit-aggrid>=0.3.4

# Development dependencies (optional)
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0